Renamed PropertyInterface interface
