+++ src/Sulu/Bundle/CategoryBundle/Content/Type/CategoryList.php
@@ -14,7 +14,7 @@ use PHPCR\NodeInterface;
 use Sulu\Bundle\CategoryBundle\Category\CategoryManagerInterface;
 use Sulu\Component\Content\Type\ComplexContentType;
 use Sulu\Component\Content\Type\ContentTypeInterface;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 
 /**
  * Content Type for the CategoryList, uses the CategoryManager-Service and the Datagrid from Husky
