+++ src/Sulu/Bundle/ContentBundle/Content/Structure/ExcerptStructureExtension.php
@@ -13,7 +13,7 @@ namespace Sulu\Bundle\ContentBundle\Content\Structure;
 use PHPCR\NodeInterface;
 use Sulu\Component\Content\Type\ContentTypeManagerInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\Extension\AbstractExtension;
 use Sulu\Component\Content\StructureInterface;
 use Sulu\Component\Content\Structure\Factory\StructureFactoryInterface;
