+++ src/Sulu/Bundle/ContentBundle/Content/Type/Checkbox.php
@@ -11,7 +11,7 @@
 namespace Sulu\Bundle\ContentBundle\Content\Type;
 
 use PHPCR\NodeInterface;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\Type\SimpleContentType;
 
 /**
