+++ src/Sulu/Bundle/ContentBundle/Tests/Unit/Content/Types/InternalLink/InternalLinksTest.php
@@ -102,7 +102,7 @@ class InternalLinksTest extends \PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
