+++ src/Sulu/Bundle/MediaBundle/Tests/Unit/MediaSelectionContentTypeTest.php
@@ -57,7 +57,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -112,7 +112,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -170,7 +170,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -221,7 +221,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -283,7 +283,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -334,7 +334,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -395,7 +395,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
@@ -446,7 +446,7 @@ class MediaSelectionContentTypeTest extends PHPUnit_Framework_TestCase
         );
 
         $property = $this->getMockForAbstractClass(
-            'Sulu\Component\Content\PropertyInterface',
+            'Sulu\Component\Content\Document\Property\PropertyInterface',
             array(),
             '',
             true,
