+++ src/Sulu/Component/Content/Mapper/Translation/MultipleTranslatedProperties.php
@@ -11,7 +11,7 @@
 namespace Sulu\Component\Content\Mapper\Translation;
 
 use Sulu\Component\Content\Property;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\Exception\NoSuchPropertyException;
 use Sulu\Component\Content\Structure;
