+++ src/Sulu/Component/Content/Mapper/Translation/TranslatedProperty.php
@@ -10,7 +10,7 @@
 
 namespace Sulu\Component\Content\Mapper\Translation;
 
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\PropertyTag;
 
 /**
@@ -20,7 +20,7 @@ use Sulu\Component\Content\PropertyTag;
 class TranslatedProperty implements PropertyInterface
 {
     /**
-     * @var \Sulu\Component\Content\PropertyInterface
+     * @var \Sulu\Component\Content\Document\Property\PropertyInterface
      */
     private $property;
 
@@ -55,7 +55,7 @@ class TranslatedProperty implements PropertyInterface
     }
 
     /**
-     * @return \Sulu\Component\Content\PropertyInterface
+     * @return \Sulu\Component\Content\Document\Property\PropertyInterface
      */
     public function getProperty()
     {
