+++ src/Sulu/Component/Content/Type/Core/ResourceLocator.php
@@ -14,7 +14,7 @@ use PHPCR\NodeInterface;
 use Sulu\Component\Content\Type\ComplexContentType;
 use Sulu\Component\Content\Type\ContentTypeInterface;
 use Sulu\Component\Content\Exception\ResourceLocatorNotFoundException;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\Type\Core\Rlp\Strategy\RLPStrategyInterface;
 
 /**
