+++ src/Sulu/Component/Util/SuluNodeHelper.php
@@ -240,10 +240,10 @@ class SuluNodeHelper
     /**
      * Return translated property
      *
-     * @param \Sulu\Component\Content\PropertyInterface $property
+     * @param \Sulu\Component\Content\Document\Property\PropertyInterface $property
      * @param string $locale
      * @param string $prefix
-     * @return \Sulu\Component\Content\PropertyInterface
+     * @return \Sulu\Component\Content\Document\Property\PropertyInterface
      */
     public function getTranslatedProperty($property, $locale, $prefix = null)
     {
