+++ tests/Sulu/Component/Content/Type/Core/BlockContentTypeTest.php
@@ -14,7 +14,7 @@ use PHPCR\NodeInterface;
 use Sulu\Component\Content\Type\ContentTypeManagerInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
 use Sulu\Component\Content\Property;
-use Sulu\Component\Content\PropertyInterface;
+use Sulu\Component\Content\Document\Property\PropertyInterface;
 use Sulu\Component\Content\Type\Core\TextArea;
 use Sulu\Component\Content\Type\Core\TextLine;
 use Sulu\Component\Content\Block\BlockProperty;
